backend/requirements.txt
fastapi
uvicorn
websockets
deepgram-sdk
openai
python-dotenv
pytest
pytest-asyncio
